Multi-tenancy plugin framework
Summary:
This patch adds a new plugin interface and a multi-tenancy module on server
side. In the first version of the plugin, the resource isolation is limited to
max connections and queries, while the APIs are generic to expand to other
resources in the future.

Main features:
  - plugin_multi_tenancy.h: defines the plugin interfaces to request and
    release resources.
  - multi-tenancy plugin will be installed and set as a global variable, so
    there will only be one multi-tenancy plugin in use on the server (similar
    to storage engine plugin). We could possible change to allow multiple
    multi-tenancy plugins in use at the same time if there is need.
  - sql_multi_tenancy.h/cc: unifies/refactors admission control into the
    multi-tenancy server module, and adds server side resource management
    interfaces.
  - resource isolation is based on connection attributes and query attributes,
    and is generic to handle per database or per user. admission control is
    modified to use query attributes.
  - connection resource management happens in two places: (1) when client
    initiates a connection, and (2) when client changes a database in the case
    of connection reuse.
  - mt_simple.h/cc: dummy multi-tenancy plugin example
  - Admission control can be set to be based on transactions (global var
    admission_control_by_trx), in addition to multi-statements.

Test Plan:
mtr and regression tests on admission control. manual test using mt_simple
plugin.
