+++ include/CMakeLists.txt
@@ -31,6 +31,7 @@ SET(HEADERS_ABI
   mysql/plugin_audit.h
   mysql/plugin_ftparser.h
   mysql/plugin_validate_password.h
+  mysql/plugin_multi_tenancy.h
 )
 
 SET(HEADERS
