+++ mysql-test/include/plugin.defs
@@ -46,3 +46,4 @@ libmemcached       plugin/innodb_memcached/daemon_memcached DAEMON_MEMCACHED dae
 innodb_engine      plugin/innodb_memcached/innodb_memcache INNODB_ENGINE
 validate_password  plugin/password_validation VALIDATE_PASSWORD validate_password
 mysql_no_login     plugin/mysql_no_login      MYSQL_NO_LOGIN    mysql_no_login
+mt_simple          plugin/mt_simple   MT_SIMPLE
