+++ mysql-test/r/mysqld--help-notwin.result
@@ -12,6 +12,8 @@ The following options may be given as the first argument:
  Option used by mysql-test for debugging and testing of
  replication.
  --admin-port=#      Port number to use for connections from admin.
+ --admission-control-by-trx 
+ Allow open transactions to go through admission control
  --admission-control-filter=name 
  Commands that are skipped in admission control checks.
  The legal values are: ALTER, BEGIN, COMMIT, CREATE,
@@ -1610,6 +1612,7 @@ The following options may be given as the first argument:
 Variables (--variable-name=value)
 abort-slave-event-count 0
 admin-port 0
+admission-control-by-trx FALSE
 admission-control-filter 
 allow-document-type FALSE
 allow-multiple-engines FALSE
