+++ mysql-test/t/admission_control_hang.test
@@ -34,7 +34,8 @@ while($i)
 }
 
 connection default;
-set @@global.admission_control_filter = 'ALTER,BEGIN,COMMIT,CREATE,DELETE,DROP,INSERT,LOAD,SELECT,SET,REPLACE,TRUNCATE,UPDATE,SHOW';
+set @@global.admission_control_filter = 'ALTER,BEGIN,COMMIT,CREATE,DELETE,DROP,INSERT,LOAD,SELECT,SET,REPLACE,TRUNCATE,UPDATE,SHOW,ROLLBACK';
+select @@global.admission_control_filter;
 
 # Verify the commands filtered above run fine.
 connection con6;
@@ -43,20 +44,56 @@ begin;
 insert into t2 values(1);
 update t2 set a=2 where a=1;
 commit;
+SHOW TABLES LIKE 't2';
+begin;
 alter table t2 rename t3;
 select * from t3;
 delete from t3;
 set @val = 1;
 truncate table t3;
+rollback;
 drop table t3;
+
+# Verify that admission_control_by_trx will honor the filters
+connection default;
+set @save_admission_control_by_trx = @@global.admission_control_by_trx;
+select @save_admission_control_by_trx;
+--echo # Turn on admission_control_by_trx
+set @@global.admission_control_by_trx = true;
+SELECT @@global.admission_control_by_trx;
+
+connection con6;
+create table t2(a int) engine=innodb;
+begin;
+insert into t2 values(1);
+update t2 set a=2 where a=1;
+commit;
 SHOW TABLES LIKE 't2';
+begin;
+alter table t2 rename t3;
+select * from t3;
+delete from t3;
+set @val = 1;
+truncate table t3;
+rollback;
+drop table t3;
 
 connection default;
-set @@global.admission_control_filter = 'ROLLBACK';
-rollback;
+set @@global.admission_control_filter = default;
+select @@global.admission_control_filter;
+
+# open transaction will be able to go through
+connection con5;
+select count(*) from t1;
+
+# restore admission_control_by_trx
+connection default;
+set @@global.admission_control_by_trx = @save_admission_control_by_trx;
+select @@global.admission_control_by_trx;
 
 connection default;
 set @@global.admission_control_filter = 'COMMIT';
+select @@global.admission_control_filter;
 
 connection con5;
 # Without the COMMIT filter set above, this query gets blocked until conflicting
