+++ mysql-test/t/mt_simple_plugin-master.opt
@@ -1,0 +1,1 @@
+$MT_SIMPLE_OPT
