+++ sql/sql_class.h
@@ -42,6 +42,7 @@
                                      THR_LOCK_INFO */
 #include "opt_trace_context.h"    /* Opt_trace_context */
 #include "rpl_gtid.h"
+#include "sql_multi_tenancy.h"
 
 #include "sql_digest_stream.h"            // sql_digest_state
 
@@ -690,6 +691,9 @@ typedef struct system_variables
 
   plugin_ref table_plugin;
   plugin_ref temp_table_plugin;
+
+  // global multi-tenancy plugin
+  plugin_ref multi_tenancy_plugin;
 
   /* Only charset part of these variables is sensible */
   const CHARSET_INFO *character_set_filesystem;
@@ -2282,41 +2286,6 @@ my_micro_time_to_timeval(ulonglong micro_time, struct timeval *tm)
 }
 
 /**
-  Per-thread information used in admission control.
-*/
-struct st_ac_node {
-#ifdef HAVE_PSI_INTERFACE
-  PSI_mutex_key key_lock;
-  PSI_cond_key key_cond;
-  PSI_mutex_info key_lock_info[1]=
-  {
-    {&key_lock, "st_ac_node::lock", 0}
-  };
-  PSI_cond_info key_cond_info[1]=
-  {
-    {&key_cond, "st_ac_node::cond", 0}
-  };
-#endif
-  mysql_mutex_t lock;
-  mysql_cond_t cond;
-  st_ac_node() {
-#ifdef HAVE_PSI_INTERFACE
-    mysql_mutex_register("sql", key_lock_info,
-                         array_elements(key_lock_info));
-    mysql_cond_register("sql", key_cond_info,
-                        array_elements(key_cond_info));
-#endif
-    mysql_mutex_init(key_lock, &lock, MY_MUTEX_INIT_FAST);
-    mysql_cond_init(key_cond, &cond, NULL);
-  }
-
-  ~st_ac_node () {
-    mysql_mutex_destroy(&lock);
-    mysql_cond_destroy(&cond);
-  }
-};
-
-/**
   @class THD
   For each client connection we create a separate thread with THD serving as
   a thread/connection descriptor
@@ -2446,6 +2415,8 @@ public:
   /* record the transaction time (including in-fly) */
   ulonglong trx_time = 0;
 
+  /* whether the session is already in admission control for queries */
+  bool is_in_ac = false;
   /**
     @note
     Some members of THD (currently 'Statement::db',
@@ -4488,6 +4459,14 @@ public:
     set_attrs_map(attrs, length, connection_attrs_map, false);
     mysql_mutex_unlock(&LOCK_thd_data);
   }
+  void update_connection_attrs(
+      const std::unordered_map<std::string, std::string> &attrs_map)
+  {
+    mysql_mutex_lock(&LOCK_thd_data);
+    for (const auto &attrs : attrs_map)
+      connection_attrs_map[attrs.first] = attrs.second;
+    mysql_mutex_unlock(&LOCK_thd_data);
+  }
 
   std::unordered_map<std::string, std::string> query_attrs_map;
   std::unordered_map<std::string, std::string> connection_attrs_map;
@@ -4614,208 +4593,6 @@ public:
   bool skip_unique_check();
   // protected by LOCK_thd_data
   std::string row_query;
-};
-
-/**
-  Class used in admission control.
-
-  Every entity (database or table or user name) will have this
-  object created and stored in the global map AC::ac_map.
-
-*/
-class Ac_info {
-  friend class AC;
-#ifdef HAVE_PSI_INTERFACE
-  PSI_mutex_key key_lock;
-  PSI_mutex_info key_lock_info[1]=
-  {
-    {&key_lock, "Ac_info::lock", 0}
-  };
-#endif
-  // Queue to track the running and waiting threads.
-  std::deque<std::shared_ptr<st_ac_node>> queue;
-  std::atomic<unsigned long> waiting_queries;
-  // Protects the queue.
-  mysql_mutex_t lock;
-public:
-  Ac_info() {
-#ifdef HAVE_PSI_INTERFACE
-    mysql_mutex_register("sql", key_lock_info,
-                         array_elements(key_lock_info));
-#endif
-    mysql_mutex_init(key_lock, &lock, MY_MUTEX_INIT_FAST);
-    waiting_queries = 0;
-  }
-  ~Ac_info() {
-    mysql_mutex_destroy(&lock);
-  }
-  // Disable copy constructor.
-  Ac_info(const Ac_info&) = delete;
-  Ac_info& operator=(const Ac_info&) = delete;
-};
-
-/**
-  Global class used to enforce per admission control limits.
-*/
-class AC {
-  // This map is protected by the rwlock LOCK_ac.
-  std::unordered_map<std::string, std::shared_ptr<Ac_info>> ac_map;
-  // Variables to track global limits
-  ulong max_running_queries, max_waiting_queries;
-  /**
-    Protects ac_map and max_running_queries/max_waiting_queries.
-
-    Locking order followed is LOCK_ac, Ac_info::lock, st_ac_node::lock.
-  */
-  mysql_rwlock_t LOCK_ac;
-#ifdef HAVE_PSI_INTERFACE
-  PSI_rwlock_key key_rwlock_LOCK_ac;
-  PSI_rwlock_info key_rwlock_LOCK_ac_info[1]=
-  {
-    {&key_rwlock_LOCK_ac, "AC::rwlock", 0}
-  };
-#endif
-
-  std::atomic<ulonglong> total_aborted_queries;
-
-public:
-  AC() {
-#ifdef HAVE_PSI_INTERFACE
-    mysql_rwlock_register("sql", key_rwlock_LOCK_ac_info,
-                          array_elements(key_rwlock_LOCK_ac_info));
-#endif
-    mysql_rwlock_init(key_rwlock_LOCK_ac, &LOCK_ac);
-    max_running_queries = 0;
-    max_waiting_queries = 0;
-    total_aborted_queries = 0;
-  }
-
-  ~AC() {
-    mysql_rwlock_destroy(&LOCK_ac);
-  }
-  // Disable copy constructor.
-  AC(const AC&) = delete;
-  AC& operator=(const AC&) = delete;
-
-  inline void signal(std::shared_ptr<st_ac_node>& ac_node) {
-    DBUG_ASSERT(ac_node && ac_node.get());
-    mysql_mutex_lock(&ac_node->lock);
-    mysql_cond_signal(&ac_node->cond);
-    mysql_mutex_unlock(&ac_node->lock);
-  }
-
-  /*
-   * Removes a dropped entity info from the global map.
-   */
-  void remove(const char* entity) {
-    std::string str(entity);
-    // First take a read lock to unblock any waiting queries.
-    mysql_rwlock_rdlock(&LOCK_ac);
-    auto it = ac_map.find(str);
-    if (it != ac_map.end()) {
-      auto ac_info  = it->second;
-      mysql_mutex_lock(&ac_info->lock);
-      while (ac_info->waiting_queries) {
-        for (uint i = max_running_queries; i < ac_info->queue.size(); ++i) {
-          signal(ac_info->queue[i]);
-        }
-      }
-      mysql_mutex_unlock(&ac_info->lock);
-    }
-    mysql_rwlock_unlock(&LOCK_ac);
-    mysql_rwlock_wrlock(&LOCK_ac);
-    it = ac_map.find(std::string(str));
-    if (it != ac_map.end()) {
-      ac_map.erase(it);
-    }
-    mysql_rwlock_unlock(&LOCK_ac);
-  }
-
-  void insert(const std::string &entity) {
-    mysql_rwlock_wrlock(&LOCK_ac);
-    if (ac_map.find(entity) == ac_map.end()) {
-      ac_map[entity] = std::make_shared<Ac_info>();
-    }
-    mysql_rwlock_unlock(&LOCK_ac);
-  }
-
-  void update_max_running_queries(ulong val) {
-    // lock to protect against erasing map iterators.
-    mysql_rwlock_wrlock(&LOCK_ac);
-    ulong old_val = max_running_queries;
-    max_running_queries = val;
-    // Signal any waiting threads which are below the new limit. Note 0 is a
-    // special case where every waiting thread needs to be signalled.
-    if (val > old_val || !val) {
-      for (auto &it: ac_map) {
-        auto &ac_info = it.second;
-        mysql_mutex_lock(&ac_info->lock);
-        for (uint i = old_val;
-             (!val || i < val) && i < ac_info->queue.size(); ++i) {
-          signal(ac_info->queue[i]);
-        }
-        mysql_mutex_unlock(&ac_info->lock);
-      }
-    }
-    mysql_rwlock_unlock(&LOCK_ac);
-  }
-
-  void update_max_waiting_queries(ulong val) {
-    mysql_rwlock_wrlock(&LOCK_ac);
-    max_waiting_queries = val;
-    mysql_rwlock_unlock(&LOCK_ac);
-  }
-
-  inline ulong get_max_running_queries() {
-    mysql_rwlock_rdlock(&LOCK_ac);
-    ulong res = max_running_queries;
-    mysql_rwlock_unlock(&LOCK_ac);
-    return res;
-  }
-
-  inline ulong get_max_waiting_queries() {
-    mysql_rwlock_rdlock(&LOCK_ac);
-    ulong res = max_waiting_queries;
-    mysql_rwlock_unlock(&LOCK_ac);
-    return res;
-  }
-
-  bool admission_control_enter(THD*, const std::string&);
-  void admission_control_exit(THD*, const std::string&);
-  void wait_for_signal(THD*, std::shared_ptr<st_ac_node>&,
-                       std::shared_ptr<Ac_info> ac_info);
-
-  ulonglong get_total_aborted_queries() const {
-    return total_aborted_queries;
-  }
-  ulong get_total_running_queries() {
-    ulonglong res= 0;
-    mysql_rwlock_rdlock(&LOCK_ac);
-    for (auto it : ac_map)
-    {
-      auto &ac_info = it.second;
-      mysql_mutex_lock(&ac_info->lock);
-      res += ac_info->queue.size() < max_running_queries ?
-        ac_info->queue.size() : max_running_queries;
-      mysql_mutex_unlock(&ac_info->lock);
-    }
-    mysql_rwlock_unlock(&LOCK_ac);
-    return res;
-  }
-  ulong get_total_waiting_queries() {
-    ulonglong res= 0;
-    mysql_rwlock_rdlock(&LOCK_ac);
-    for (auto it : ac_map)
-    {
-      auto &ac_info = it.second;
-      mysql_mutex_lock(&ac_info->lock);
-      if (ac_info->queue.size() > max_running_queries)
-        res += ac_info->queue.size() - max_running_queries;
-      mysql_mutex_unlock(&ac_info->lock);
-    }
-    mysql_rwlock_unlock(&LOCK_ac);
-    return res;
-  }
 };
 
 /*
